Refactor `models.py` into a package
