+++ core/models/__init__.py
@@ -1,0 +1,10 @@
+from .event import Event, EventPageContent, EventPageMenu
+from .user import User, user_invite
+
+__all__ = [
+    'Event',
+    'EventPageContent',
+    'EventPageMenu',
+    'User',
+    'user_invite'
+]
