+++ core/models/managers/__init__.py
@@ -1,0 +1,2 @@
+from .event import EventManager
+from .user import UserManager
