+++ core/models/managers/user.py
@@ -1,0 +1,18 @@
+from django.contrib.auth import models as auth_models
+
+
+class UserManager(auth_models.BaseUserManager):
+
+    def create_user(self, email, password=None):
+        if not email:
+            raise ValueError("Users must have an email address")
+        user = self.model(email=self.normalize_email(email))
+        user.set_password(password)
+        user.save(using=self._db)
+        return user
+
+    def create_superuser(self, email, password):
+        user = self.create_user(email, password=password)
+        user.is_superuser = user.is_staff = True
+        user.save(using=self._db)
+        return user
